Fixed an issue with the history list
